# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handler.SetRequest:
    properties:
      value: {}
    required:
    - value
    type: object
...
  response.DataResponse-array_team_SettingsItemDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据
        items:
          $ref: '#/definitions/team.SettingsItemDTO'
        type: array
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-team_SettingsItemDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/team.SettingsItemDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  setting.DependsOnConfigDTO:
    properties:
      key:
        type: string
      operator:
        description: eq, ne, gt, lt（默认 eq）
        type: string
      value: {}
    type: object
...
  setting.SelectOptionDTO:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
...
  setting.UIConfigDTO:
    properties:
      depends_on:
        allOf:
        - $ref: '#/definitions/setting.DependsOnConfigDTO'
        description: 依赖关系
      hint:
        description: 输入提示
        type: string
      options:
        description: 下拉选项
        items:
          $ref: '#/definitions/setting.SelectOptionDTO'
        type: array
    type: object
  team.SettingsItemDTO:
    properties:
      category:
        description: 分类 Key（前端按需分组）
        type: string
      configurable_at:
        description: 最大可配置级别
        type: string
      default_value:
        description: 系统默认值
      group:
        description: 分组名称
        type: string
      input_type:
        description: 控件类型
        type: string
      is_customized:
        description: 是否用户自定义
        type: boolean
      key:
        type: string
      label:
        type: string
      order:
        type: integer
      ui_config:
        allOf:
        - $ref: '#/definitions/setting.UIConfigDTO'
        description: hint/options/depends_on
      validation:
        description: JSON Logic 规则
      value:
        description: 实际生效值
      value_type:
        type: string
      visible_at:
        description: 最小可见级别
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/lwmacct/260103-ddd-iam-bc
  description: 基于 DDD + CQRS 架构的可复用模块库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go DDD Package Library API
  version: "1.0"
paths:
...
  /api/org/{org_id}/teams/{team_id}/settings:
    get:
      consumes:
      - application/json
      description: 获取当前团队的配置列表（支持三级继承：团队>组织>系统默认值）
      parameters:
      - description: 组织ID
        in: path
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: 可选：按分类 Key 过滤（如 "general"）
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 配置列表
          schema:
            $ref: '#/definitions/response.DataResponse-array_team_SettingsItemDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 团队配置列表
      tags:
      - settings-team
  /api/org/{org_id}/teams/{team_id}/settings/{key}:
    delete:
      consumes:
      - application/json
      description: 重置指定配置项（删除团队自定义值，恢复组织配置或系统默认值）
      parameters:
      - description: 组织ID
        in: path
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: 配置键名
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 重置团队配置
      tags:
      - settings-team
    get:
      consumes:
      - application/json
      description: 获取指定配置项的值（支持三级继承：团队>组织>系统默认值）
      parameters:
      - description: 组织ID
        in: path
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: 配置键名
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 配置信息
          schema:
            $ref: '#/definitions/response.DataResponse-team_SettingsItemDTO'
        "400":
          description: 无效的键名
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取团队配置
      tags:
      - settings-team
    put:
      consumes:
      - application/json
      description: 设置指定配置项的值（团队自定义覆盖）
      parameters:
      - description: 组织ID
        in: path
        name: org_id
        required: true
        type: integer
      - description: 团队ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: 配置键名
        in: path
        name: key
        required: true
        type: string
      - description: 设置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/response.DataResponse-team_SettingsItemDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置团队配置
      tags:
      - settings-team
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"